Changed a whole bunch of things.
Not pleased with the changes I made first, I have removed some of
them to make new and better one.

Electives are back, along with debates and Torturing.

There was to many place where the grammar was wrong, so I will go
through again, and change it manually (as appose to lots of search
replace).

New things are now:

Giants are now Biggly.
Dwarfs are dorfs, short sturdy creatures fond of drink and industry.
Incestues is now keeing it  in the family.
Men-at-Arms are Men-with-Arms.